Bug 10995: Make inner window a multiple of 200x100
In order to be sure that no other extension is resizing the inner window
due to merging its overlay into the chrome we resize the window (again)
with setTimeout(0). This minimizes the usability issues as well as the
window just gets larger by some pixels while being visible (if it gets
larger at all, i.e. if proper resizing before the window got visible did
not succeed).